lib/davis/python/requirements.txt
Cython==0.24
Pillow==3.2.0
PyYAML==3.11
argparse==1.2.1
cycler==0.10.0
dask==0.8.1
decorator==4.0.9
easydict==1.6
future==0.15.2
h5py==2.6.0
matplotlib==1.5.1
networkx==1.11
numpy==1.11.0
prettytable==0.7.2
pyparsing==2.1.1
python-dateutil==2.5.2
pytz==2016.3
scikit-image==0.12.3
scipy==0.17.0
six==1.10.0
toolz==0.7.4
wsgiref==0.1.2
